[sync] Bug 1521864 -  It should be possible to open the debugger in a tab. r=davidwalsh
