+++ webpack.config.js
@@ -28,7 +28,7 @@ const webpackConfig = {
   entry: {
     // We always generate the debugger bundle, but we will only copy the CSS
     // artifact over to mozilla-central.
-    debugger: getEntry("src/main.js"),
+    debugger: getEntry(isProduction ? "src/main.js" : "src/main.development.js"),
     "parser-worker": getEntry("src/workers/parser/worker.js"),
     "pretty-print-worker": getEntry("src/workers/pretty-print/worker.js"),
     "search-worker": getEntry("src/workers/search/worker.js"),
